+++ encodings/alp/src/alp/compute/take.rs
@@ -1,0 +1,30 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+use vortex_array::compute::{TakeKernel, TakeKernelAdapter, take};
+use vortex_array::{Array, ArrayRef, register_kernel};
+use vortex_error::VortexResult;
+
+use crate::{ALPArray, ALPVTable};
+
+impl TakeKernel for ALPVTable {
+    fn take(&self, array: &ALPArray, indices: &dyn Array) -> VortexResult<ArrayRef> {
+        let taken_encoded = take(array.encoded(), indices)?;
+        let taken_patches = array
+            .patches()
+            .map(|p| p.take(indices))
+            .transpose()?
+            .flatten()
+            .map(|patches| {
+                patches.cast_values(
+                    &array
+                        .dtype()
+                        .with_nullability(taken_encoded.dtype().nullability()),
+                )
+            })
+            .transpose()?;
+        Ok(ALPArray::try_new(taken_encoded, array.exponents(), taken_patches)?.to_array())
+    }
+}
+
+register_kernel!(TakeKernelAdapter(ALPVTable).lift());
