+++ encodings/fastlanes/src/for/compute/mod.rs
@@ -73,12 +73,12 @@ mod test {
         let values = PrimitiveArray::from_iter([100i32, 101, 102, 103, 104]);
         let reference = Scalar::from(100i32);
         let for_array = FoRArray::try_new(values.as_ref(), reference).unwrap();
-        test_numeric(for_array.as_ref(), for_array.as_ref());
+        test_numeric(for_array.into_array());
 
         // Test with different array
         let values2 = PrimitiveArray::from_iter([200i32, 201, 202, 203, 204]);
         let reference2 = Scalar::from(200i32);
         let for_array2 = FoRArray::try_new(values2.as_ref(), reference2).unwrap();
-        test_numeric(for_array.as_ref(), for_array2.as_ref());
+        test_numeric(for_array2.into_array());
     }
 }
